refactor(sample): Signal sample 改用 Float32 reference path

+++ samples/OpenDeviceStudio.Sample.DataAcquisition/Acquisition/SampleFrame.cs
@@ -3,4 +3,4 @@ namespace OpenDeviceStudio.Sample.DataAcquisition.Acquisition;
 public sealed record SampleFrame(
     long Sequence,
     DateTimeOffset Timestamp,
-    double[] Channels);
+    float[] Channels);
